Lass route 102 and raltz fix
The lass and the Zigzagoons on route 102 have been fixed. The zigzagoons shouldn't appear until the lass quest is started and the lass quest switch should now active once you talk to her to get the quest.

Also, Ralts in the first two encounters is normal and on route 103 she has her new form!