Pokemon fan club fix (again), Growlithe fix!
So when you start the berry quest "Berry Bundle for the Club" for the Pokemon fan club, when you go to the berry shop daisy wouldn't be there. She is now there to push the quest forward!

You where able to interact with the "Find Hatenna" quest (on route 111) without the proper trigger. Now it should be fix.

With starter growlithe their first event in the petalburg woods after you get the first badge had a odd trigger, that should be fixed!

All substarters should be in their correct areas!