Less work for MDH
Goodra can now be Breed and give an Egg.
Greninja can now be Breed.
Some fixes with some Breeding scenes that were needed.
the "me" has been change to "be"
Goodra and Greninja added to the Gallery and fully functional in there.

Added the Day icon for Whimsicott, Greninja and Goodra.

Made the Impregnation and Breeding common event for Goodra and Greninja

Fixed Goodra's wrong text in the Sign of love quest