the Basement Veevee's Plus some Fixes
Fixed things with Lilith name, request names and Showing a fighting type quest text. (if you already have the quest it won't change the text. The pokemon you need to show is Machamp)

Made the Eevee[2] and Eevee[3] basement be fully functional.

New switches and Variables added (So be careful for your switches to not get erased MDH)

Changed a all the eeveelution breeding scenes to work properly, added an cutscene event for eevee[2] in the Cutscene room.

Changed the impregnation common event to work with all the Eevees now. changed the Day icon event to only display eevee (If people don't get it across their minds i'll blow up)