Maractus and FlygonSG
Maracter and SG flygon fights are ready I think. After you've defeated Inesse and talk to Houndoom, go to gals house and you'll encounter Inesse. Go back to your house and you should be able to see Zinnia. The quest isn't in the log yet! But it will be when I'm not too tired.