Lopunny and bike update!
Lopunny's fight should be in the game now! As well as the grand tour quest line being updated to reflect this. Lopunny and Inesse should appear in your home once they are defeated. Voidkip should now appear infront of your home with the bike when you defeat Inesse and houndoom!